server/requirements.txt
fastapi[standard]==0.115.2
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
jwt==1.3.1
alembic==1.13.3
fastapi-mail==1.4.1
Jinja2==3.1.4
redis==5.1.1
celery[librabbitmq,redis,auth,msgpack]==5.4.0
asgiref==3.8.1
flower==2.0.1
python-dotenv==1.0.1
psycopg2==2.9.10
stripe==11.3.0
google-auth
requests